moved notifications from presentation layer
