+++ myProject/app/build.gradle
@@ -54,9 +54,9 @@ dependencies {
 
     implementation 'androidx.core:core-ktx:1.7.0'
     implementation 'androidx.appcompat:appcompat:1.5.1'
-    implementation project(path: ':domain')
 
-    //implementation project(path: ':data')
+    implementation project(path: ':domain')
+    implementation project(path: ':data')
 
     implementation 'com.google.android.material:material:1.8.0-alpha03'
     implementation 'androidx.constraintlayout:constraintlayout:2.1.4'
