+++ myProject/app/src/main/java/com/blblblbl/myapplication/di/DataModule.kt
@@ -3,7 +3,6 @@ package com.blblblbl.myapplication.di
 import com.blblblbl.myapplication.data.DownloadNotifications
 import com.blblblbl.myapplication.data.repository.RepositoryImpl
 import com.blblblbl.myapplication.domain.repository.Repository
-import com.blblblbl.myapplication.presentation.viewModel.DownloadNotificationsImpl
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
